Provide a usage run through in the README
To help people who are newer to elm and unsure of how to interact with
libraries. In particular I wanted to cover the 'how to upate a date
picker' case which is achieved by creating a new date picker with the
new date which isn't necessarily a pattern that beginner's will think
of.